backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",
    "greenlet>=3.0.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
    "celery>=5.3.0",
    "redis>=4.6.0,<6.0",
    "alembic>=1.13.0"
]
